# swagger-static/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /api/v1/projects/vote/{id}:
    put:
      operationId: ProjectsController_voteOneById
      summary: Голосование за проект по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_0
      security: *ref_1
...
  /api/v1/votes:
    post:
      operationId: VotesController_createOne
      summary: Создание нового голоса
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVoteDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
        '401':
          description: Unauthorized
      tags: &ref_4
        - votes
      security: &ref_5
        - bearer: []
    get:
      operationId: VotesController_findAll
      summary: Получение списка голосов
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
        '401':
          description: Unauthorized
      tags: *ref_4
      security: *ref_5
  /api/v1/votes/{id}:
    get:
      operationId: VotesController_findOneById
      summary: Получение голоса по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_4
      security: *ref_5
    put:
      operationId: VotesController_updateOneById
      summary: Обновление голоса по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVoteDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_4
      security: *ref_5
    delete:
      operationId: VotesController_deleteOneById
      summary: Удаление голоса по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vote'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_4
      security: *ref_5
...
info:
  title: Студенческий код
  description: Это открытое API для веб-приложения "Студенческий код".
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ProjectLinksDto:
      type: object
      properties:
        main:
          type: string
          description: Основная ссылка
        github:
          type: string
          description: Ссылка на GitHub
        demo:
          type: string
          description: Ссылка на демо
      required:
        - main
        - github
        - demo
...
    ObjectId:
      type: object
      properties: {}
    TagNameDto:
      type: object
      properties:
        en:
          type: string
          description: Название на английском
        ru:
          type: string
          description: Название на русском
      required:
        - en
        - ru
    Tag:
      type: object
      properties:
        _id:
          description: Идентификатор
          allOf:
            - $ref: '#/components/schemas/ObjectId'
        name:
          description: Название
          allOf:
            - $ref: '#/components/schemas/TagNameDto'
        icon:
          type: string
          description: Ссылка на иконку
        slug:
          type: string
          description: Ключевое слово
      required:
        - _id
        - name
        - icon
        - slug
    UserFullNameDto:
      type: object
      properties:
        surname:
          type: string
          description: Фамилия
        name:
          type: string
          description: Имя
        patronymic:
          type: string
          description: Отчество
      required:
        - surname
        - name
        - patronymic
    UserLinksDto:
      type: object
      properties:
        github:
          type: string
          description: Гитхаб
        vkontakte:
          type: string
          description: Вконтакте
        telegram:
          type: string
          description: Телеграм
      required:
        - github
        - vkontakte
        - telegram
    AwardNameDto:
      type: object
      properties:
        en:
          type: string
          description: Название на английском
        ru:
          type: string
          description: Название на русском
      required:
        - en
        - ru
    Award:
      type: object
      properties:
        _id:
          description: Идентификатор
          allOf:
            - $ref: '#/components/schemas/ObjectId'
        name:
          description: Название
          allOf:
            - $ref: '#/components/schemas/AwardNameDto'
        icon:
          type: string
          description: Ссылка на иконку
      required:
        - _id
        - name
        - icon
    User:
      type: object
      properties:
        _id:
          description: Идентификатор
          allOf:
            - $ref: '#/components/schemas/ObjectId'
        username:
          type: string
          description: Имя пользователя
        email:
          type: string
          description: Электронная почта
        password:
          type: string
          description: Пароль
        role:
          type: string
          description: Роль
          enum:
            - user
            - admin
        refresh_token:
          type: string
          description: Токен обновления
        full_name:
          description: ФИО
          allOf:
            - $ref: '#/components/schemas/UserFullNameDto'
        avatar:
          type: string
          description: Ссылка на аватар
        about:
          type: string
          description: О себе
        links:
          description: Ссылки
          allOf:
            - $ref: '#/components/schemas/UserLinksDto'
        awards:
          description: Награды
          type: array
          items:
            $ref: '#/components/schemas/Award'
        projects:
          description: Проекты
          type: array
          items:
            type: array
      required:
        - _id
        - username
        - email
        - password
        - role
        - refresh_token
        - full_name
        - avatar
        - about
        - links
        - awards
        - projects
    Project:
      type: object
      properties:
        _id:
          description: Идентификатор
          allOf:
            - $ref: '#/components/schemas/ObjectId'
        title:
          type: string
          description: Название
        tagline:
          type: string
          description: Слоган
        status:
          type: string
          description: Статус
          enum:
            - draft
            - published
            - archived
        description:
          type: string
          description: Описание
        flames:
          type: number
          description: Количество огоньков
        links:
          description: Ссылки
          allOf:
            - $ref: '#/components/schemas/ProjectLinksDto'
        logo:
          type: string
          description: Ссылка на логотип
        screenshots:
          description: Массив скриншотов
          type: array
          items:
            type: string
        price:
          type: string
          description: Цена
          enum:
            - free
            - free_options
            - payment_required
        rating:
          type: number
          description: Рейтинг
        tags:
          description: Теги
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        creator:
          description: Создатель
          allOf:
            - $ref: '#/components/schemas/User'
      required:
        - _id
        - title
        - tagline
        - status
        - description
        - flames
        - links
        - logo
        - screenshots
        - price
        - rating
        - tags
        - creator
...
    CreateVoteDto:
      type: object
      properties:
        project:
          type: string
          description: Проект
        voter:
          type: string
          description: Голосующий
      required:
        - project
        - voter
    Vote:
      type: object
      properties:
        _id:
          description: Идентификатор
          allOf:
            - $ref: '#/components/schemas/ObjectId'
        project:
          description: Проект
          allOf:
            - $ref: '#/components/schemas/Project'
        voter:
          description: Голосующий
          allOf:
            - $ref: '#/components/schemas/User'
      required:
        - _id
        - project
        - voter
    UpdateVoteDto:
      type: object
      properties: {}
...